# frontend/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ConciliacionArchivosSelection": {
        "properties": {
          "archivo_a_id": {
            "exclusiveMinimum": 0,
            "title": "Archivo A Id",
            "type": "integer"
          },
          "archivo_b_id": {
            "exclusiveMinimum": 0,
            "title": "Archivo B Id",
            "type": "integer"
          }
        },
        "required": [
          "archivo_a_id",
          "archivo_b_id"
        ],
        "title": "ConciliacionArchivosSelection",
        "type": "object"
      },
...
      "ConciliacionResumenRead": {
        "properties": {
          "conciliados": {
            "title": "Conciliados",
            "type": "integer"
          },
          "diferencias_importe": {
            "title": "Diferencias Importe",
            "type": "integer"
          },
          "duplicados_archivo_a": {
            "title": "Duplicados Archivo A",
            "type": "integer"
          },
          "duplicados_archivo_b": {
            "title": "Duplicados Archivo B",
            "type": "integer"
          },
          "ejecucion_id": {
            "title": "Ejecucion Id",
            "type": "integer"
          },
          "errores_formato": {
            "title": "Errores Formato",
            "type": "integer"
          },
          "estado_ejecucion": {
            "title": "Estado Ejecucion",
            "type": "string"
          },
          "requiere_revision": {
            "title": "Requiere Revision",
            "type": "integer"
          },
          "solo_archivo_a": {
            "title": "Solo Archivo A",
            "type": "integer"
          },
          "solo_archivo_b": {
            "title": "Solo Archivo B",
            "type": "integer"
          },
          "total_resultados": {
            "title": "Total Resultados",
            "type": "integer"
          }
        },
        "required": [
          "ejecucion_id",
          "total_resultados",
          "conciliados",
          "diferencias_importe",
          "solo_archivo_a",
          "solo_archivo_b",
          "duplicados_archivo_a",
          "duplicados_archivo_b",
          "errores_formato",
          "requiere_revision",
          "estado_ejecucion"
        ],
        "title": "ConciliacionResumenRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RechazarEjecucionRequest": {
        "additionalProperties": false,
        "properties": {
          "motivo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Motivo"
          }
        },
        "title": "RechazarEjecucionRequest",
        "type": "object"
      },
...
      "ResultadoConciliacionRead": {
        "properties": {
          "clave_referencia": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clave Referencia"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "datos_archivo_a_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datos Archivo A Json"
          },
          "datos_archivo_b_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datos Archivo B Json"
          },
          "diferencia_importe": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diferencia Importe"
          },
          "ejecucion_id": {
            "title": "Ejecucion Id",
            "type": "integer"
          },
          "estado_resultado": {
            "title": "Estado Resultado",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "observacion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observacion"
          },
          "requiere_revision": {
            "title": "Requiere Revision",
            "type": "boolean"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "id",
          "ejecucion_id",
          "clave_referencia",
          "estado_resultado",
          "datos_archivo_a_json",
          "datos_archivo_b_json",
          "diferencia_importe",
          "requiere_revision",
          "observacion",
          "created_at",
          "updated_at"
        ],
        "title": "ResultadoConciliacionRead",
        "type": "object"
      },
      "ResultadoRevisionUpdate": {
        "additionalProperties": false,
        "properties": {
          "expected_updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Updated At"
          },
          "observacion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observacion"
          },
          "requiere_revision": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requiere Revision"
          }
        },
        "required": [
          "expected_updated_at"
        ],
        "title": "ResultadoRevisionUpdate",
        "type": "object"
      },
      "RevisionResumenRead": {
        "properties": {
          "conciliados": {
            "title": "Conciliados",
            "type": "integer"
          },
          "diferencias_importe": {
            "title": "Diferencias Importe",
            "type": "integer"
          },
          "duplicados_archivo_a": {
            "title": "Duplicados Archivo A",
            "type": "integer"
          },
          "duplicados_archivo_b": {
            "title": "Duplicados Archivo B",
            "type": "integer"
          },
          "ejecucion_id": {
            "title": "Ejecucion Id",
            "type": "integer"
          },
          "errores_formato": {
            "title": "Errores Formato",
            "type": "integer"
          },
          "estado_ejecucion": {
            "title": "Estado Ejecucion",
            "type": "string"
          },
          "pendientes_revision": {
            "title": "Pendientes Revision",
            "type": "integer"
          },
          "revisados": {
            "title": "Revisados",
            "type": "integer"
          },
          "solo_archivo_a": {
            "title": "Solo Archivo A",
            "type": "integer"
          },
          "solo_archivo_b": {
            "title": "Solo Archivo B",
            "type": "integer"
          },
          "total_resultados": {
            "title": "Total Resultados",
            "type": "integer"
          }
        },
        "required": [
          "ejecucion_id",
          "estado_ejecucion",
          "total_resultados",
          "pendientes_revision",
          "revisados",
          "conciliados",
          "diferencias_importe",
          "solo_archivo_a",
          "solo_archivo_b",
          "duplicados_archivo_a",
          "duplicados_archivo_b",
          "errores_formato"
        ],
        "title": "RevisionResumenRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Automatizador Administrativo Web",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/conciliaciones/resultados/{resultado_id}/revision": {
      "patch": {
        "operationId": "update_revision_conciliaciones_resultados__resultado_id__revision_patch",
        "parameters": [
          {
            "in": "path",
            "name": "resultado_id",
            "required": true,
            "schema": {
              "title": "Resultado Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResultadoRevisionUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultadoConciliacionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Revision",
        "tags": [
          "Conciliaciones"
        ]
      }
    },
    "/conciliaciones/{ejecucion_id}/aprobar": {
      "post": {
        "operationId": "approve_reconciliation_conciliaciones__ejecucion_id__aprobar_post",
        "parameters": [
          {
            "in": "path",
            "name": "ejecucion_id",
            "required": true,
            "schema": {
              "title": "Ejecucion Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionResumenRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Approve Reconciliation",
        "tags": [
          "Conciliaciones"
        ]
      }
    },
    "/conciliaciones/{ejecucion_id}/archivos": {
      "get": {
        "operationId": "read_selected_files_conciliaciones__ejecucion_id__archivos_get",
        "parameters": [
          {
            "in": "path",
            "name": "ejecucion_id",
            "required": true,
            "schema": {
              "title": "Ejecucion Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConciliacionArchivosSelection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read Selected Files",
        "tags": [
          "Conciliaciones"
        ]
      },
      "put": {
        "operationId": "replace_selected_files_conciliaciones__ejecucion_id__archivos_put",
        "parameters": [
          {
            "in": "path",
            "name": "ejecucion_id",
            "required": true,
            "schema": {
              "title": "Ejecucion Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConciliacionArchivosSelection"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConciliacionArchivosSelection"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Replace Selected Files",
        "tags": [
          "Conciliaciones"
        ]
      }
    },
    "/conciliaciones/{ejecucion_id}/ejecutar": {
      "post": {
        "operationId": "execute_mapping_conciliaciones__ejecucion_id__ejecutar_post",
        "parameters": [
          {
            "in": "path",
            "name": "ejecucion_id",
            "required": true,
            "schema": {
              "title": "Ejecucion Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConciliacionResumenRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Execute Mapping",
        "tags": [
          "Conciliaciones"
        ]
      }
    },
    "/conciliaciones/{ejecucion_id}/exportar": {
      "get": {
        "operationId": "export_results_conciliaciones__ejecucion_id__exportar_get",
        "parameters": [
          {
            "in": "path",
            "name": "ejecucion_id",
            "required": true,
            "schema": {
              "title": "Ejecucion Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Archivo XLSX con los resultados de la conciliación"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Export Results",
        "tags": [
          "Conciliaciones"
        ]
      }
    },
...
    "/conciliaciones/{ejecucion_id}/rechazar": {
      "post": {
        "operationId": "reject_reconciliation_conciliaciones__ejecucion_id__rechazar_post",
        "parameters": [
          {
            "in": "path",
            "name": "ejecucion_id",
            "required": true,
            "schema": {
              "title": "Ejecucion Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RechazarEjecucionRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Rechazo In"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionResumenRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Reject Reconciliation",
        "tags": [
          "Conciliaciones"
        ]
      }
    },
    "/conciliaciones/{ejecucion_id}/resultados": {
      "get": {
        "operationId": "read_results_conciliaciones__ejecucion_id__resultados_get",
        "parameters": [
          {
            "in": "path",
            "name": "ejecucion_id",
            "required": true,
            "schema": {
              "title": "Ejecucion Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "estado_resultado",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Estado Resultado"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ResultadoConciliacionRead"
                  },
                  "title": "Response Read Results Conciliaciones  Ejecucion Id  Resultados Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read Results",
        "tags": [
          "Conciliaciones"
        ]
      }
    },
    "/conciliaciones/{ejecucion_id}/revision-resumen": {
      "get": {
        "operationId": "read_revision_summary_conciliaciones__ejecucion_id__revision_resumen_get",
        "parameters": [
          {
            "in": "path",
            "name": "ejecucion_id",
            "required": true,
            "schema": {
              "title": "Ejecucion Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionResumenRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read Revision Summary",
        "tags": [
          "Conciliaciones"
        ]
      }
    },
...
  }
}